requirements.txt
streamlit==1.31.0
supabase==1.2.0
anthropic==0.8.0
python-dotenv==1.0.0
pillow==10.0.0